добавляет разбивку на модули
