+++ index.html
@@ -234,7 +234,7 @@
 
   <!-- Основные скрипты -->
 
-  <script src="./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 
 </body>
 </html>
